typecast fixes, recursion

+++ src/promote.c
@@ -9,32 +9,32 @@ switch(a->type)
 case I8: 
     switch(b->type)
     {
-        case  I8: a->poly.i8 = a->poly.i8 OP b->poly.i8;  a->type =  I8; break;
-        case  U8: a->poly.i8 = a->poly.i8 OP b->poly.u8;  a->type =  U8; break;
-        case I16: a->poly.i8 = a->poly.i8 OP b->poly.i16; a->type = I16; break;
-        case U16: a->poly.i8 = a->poly.i8 OP b->poly.u16; a->type = U16; break;
-        case I32: a->poly.i8 = a->poly.i8 OP b->poly.i32; a->type = I32; break;
-        case U32: a->poly.i8 = a->poly.i8 OP b->poly.u32; a->type = U32; break;
-        case I64: a->poly.i8 = a->poly.i8 OP b->poly.i64; a->type = I64; break;
-        case U64: a->poly.i8 = a->poly.i8 OP b->poly.u64; a->type = U64; break;
-        case F32: a->poly.i8 = a->poly.i8 OP b->poly.f32; a->type = F32; break;
-        case F64: a->poly.i8 = a->poly.i8 OP b->poly.f64; a->type = F64; break;
+        case  I8: a->poly. i8 = a->poly.i8 OP b->poly.i8;  a->type =  I8; break;
+        case  U8: a->poly. u8 = a->poly.i8 OP b->poly.u8;  a->type =  U8; break;
+        case I16: a->poly.i16 = a->poly.i8 OP b->poly.i16; a->type = I16; break;
+        case U16: a->poly.u16 = a->poly.i8 OP b->poly.u16; a->type = U16; break;
+        case I32: a->poly.i32 = a->poly.i8 OP b->poly.i32; a->type = I32; break;
+        case U32: a->poly.u32 = a->poly.i8 OP b->poly.u32; a->type = U32; break;
+        case I64: a->poly.i64 = a->poly.i8 OP b->poly.i64; a->type = I64; break;
+        case U64: a->poly.u64 = a->poly.i8 OP b->poly.u64; a->type = U64; break;
+        case F32: a->poly.f32 = a->poly.i8 OP b->poly.f32; a->type = F32; break;
+        case F64: a->poly.f64 = a->poly.i8 OP b->poly.f64; a->type = F64; break;
         default: BAIL; break;
     }
     break;
 case U8: 
     switch(b->type)
     {
-        case  I8: a->poly.u8 = a->poly.u8 OP b->poly.i8;  a->type =  U8; break;
-        case  U8: a->poly.u8 = a->poly.u8 OP b->poly.u8;  a->type =  U8; break;
-        case I16: a->poly.u8 = a->poly.u8 OP b->poly.i16; a->type = I16; break;
-        case U16: a->poly.u8 = a->poly.u8 OP b->poly.u16; a->type = U16; break;
-        case I32: a->poly.u8 = a->poly.u8 OP b->poly.i32; a->type = I32; break;
-        case U32: a->poly.u8 = a->poly.u8 OP b->poly.u32; a->type = U32; break;
-        case I64: a->poly.u8 = a->poly.u8 OP b->poly.i64; a->type = I64; break;
-        case U64: a->poly.u8 = a->poly.u8 OP b->poly.u64; a->type = U64; break;
-        case F32: a->poly.u8 = a->poly.u8 OP b->poly.f32; a->type = F32; break;
-        case F64: a->poly.u8 = a->poly.u8 OP b->poly.f64; a->type = F64; break;
+        case  I8: a->poly. u8 = a->poly.u8 OP b->poly.i8;  a->type =  U8; break;
+        case  U8: a->poly. u8 = a->poly.u8 OP b->poly.u8;  a->type =  U8; break;
+        case I16: a->poly.i16 = a->poly.u8 OP b->poly.i16; a->type = I16; break;
+        case U16: a->poly.u16 = a->poly.u8 OP b->poly.u16; a->type = U16; break;
+        case I32: a->poly.i32 = a->poly.u8 OP b->poly.i32; a->type = I32; break;
+        case U32: a->poly.u32 = a->poly.u8 OP b->poly.u32; a->type = U32; break;
+        case I64: a->poly.i64 = a->poly.u8 OP b->poly.i64; a->type = I64; break;
+        case U64: a->poly.u64 = a->poly.u8 OP b->poly.u64; a->type = U64; break;
+        case F32: a->poly.f32 = a->poly.u8 OP b->poly.f32; a->type = F32; break;
+        case F64: a->poly.f64 = a->poly.u8 OP b->poly.f64; a->type = F64; break;
         default: BAIL; break;
     }
     break;
@@ -44,13 +44,13 @@ case I16:
         case  I8: a->poly.i16 = a->poly.i16 OP b->poly.i8;  a->type = I16; break;
         case  U8: a->poly.i16 = a->poly.i16 OP b->poly.u8;  a->type = I16; break;
         case I16: a->poly.i16 = a->poly.i16 OP b->poly.i16; a->type = I16; break;
-        case U16: a->poly.i16 = a->poly.i16 OP b->poly.u16; a->type = U16; break;
-        case I32: a->poly.i16 = a->poly.i16 OP b->poly.i32; a->type = I32; break;
-        case U32: a->poly.i16 = a->poly.i16 OP b->poly.u32; a->type = U32; break;
-        case I64: a->poly.i16 = a->poly.i16 OP b->poly.i64; a->type = I64; break;
-        case U64: a->poly.i16 = a->poly.i16 OP b->poly.u64; a->type = U64; break;
-        case F32: a->poly.i16 = a->poly.i16 OP b->poly.f32; a->type = F32; break;
-        case F64: a->poly.i16 = a->poly.i16 OP b->poly.f64; a->type = F64; break;
+        case U16: a->poly.u16 = a->poly.i16 OP b->poly.u16; a->type = U16; break;
+        case I32: a->poly.i32 = a->poly.i16 OP b->poly.i32; a->type = I32; break;
+        case U32: a->poly.u32 = a->poly.i16 OP b->poly.u32; a->type = U32; break;
+        case I64: a->poly.i64 = a->poly.i16 OP b->poly.i64; a->type = I64; break;
+        case U64: a->poly.u64 = a->poly.i16 OP b->poly.u64; a->type = U64; break;
+        case F32: a->poly.f32 = a->poly.i16 OP b->poly.f32; a->type = F32; break;
+        case F64: a->poly.f64 = a->poly.i16 OP b->poly.f64; a->type = F64; break;
         default: BAIL; break;
     }
     break;
@@ -61,12 +61,12 @@ case U16:
         case  U8: a->poly.u16 = a->poly.u16 OP b->poly.u8;  a->type = U16; break;
         case I16: a->poly.u16 = a->poly.u16 OP b->poly.i16; a->type = U16; break;
         case U16: a->poly.u16 = a->poly.u16 OP b->poly.u16; a->type = U16; break;
-        case I32: a->poly.u16 = a->poly.u16 OP b->poly.i32; a->type = I32; break;
-        case U32: a->poly.u16 = a->poly.u16 OP b->poly.u32; a->type = U32; break;
-        case I64: a->poly.u16 = a->poly.u16 OP b->poly.i64; a->type = I64; break;
-        case U64: a->poly.u16 = a->poly.u16 OP b->poly.u64; a->type = U64; break;
-        case F32: a->poly.u16 = a->poly.u16 OP b->poly.f32; a->type = F32; break;
-        case F64: a->poly.u16 = a->poly.u16 OP b->poly.f64; a->type = F64; break;
+        case I32: a->poly.i32 = a->poly.u16 OP b->poly.i32; a->type = I32; break;
+        case U32: a->poly.u32 = a->poly.u16 OP b->poly.u32; a->type = U32; break;
+        case I64: a->poly.i64 = a->poly.u16 OP b->poly.i64; a->type = I64; break;
+        case U64: a->poly.u64 = a->poly.u16 OP b->poly.u64; a->type = U64; break;
+        case F32: a->poly.f32 = a->poly.u16 OP b->poly.f32; a->type = F32; break;
+        case F64: a->poly.f64 = a->poly.u16 OP b->poly.f64; a->type = F64; break;
         default: BAIL; break;
     }
     break;
@@ -78,11 +78,11 @@ case I32:
         case I16: a->poly.i32 = a->poly.i32 OP b->poly.i16; a->type = I32; break;
         case U16: a->poly.i32 = a->poly.i32 OP b->poly.u16; a->type = I32; break;
         case I32: a->poly.i32 = a->poly.i32 OP b->poly.i32; a->type = I32; break;
-        case U32: a->poly.i32 = a->poly.i32 OP b->poly.u32; a->type = U32; break;
-        case I64: a->poly.i32 = a->poly.i32 OP b->poly.i64; a->type = I64; break;
-        case U64: a->poly.i32 = a->poly.i32 OP b->poly.u64; a->type = U64; break;
-        case F32: a->poly.i32 = a->poly.i32 OP b->poly.f32; a->type = F32; break;
-        case F64: a->poly.i32 = a->poly.i32 OP b->poly.f64; a->type = F64; break;
+        case U32: a->poly.u32 = a->poly.i32 OP b->poly.u32; a->type = U32; break;
+        case I64: a->poly.i64 = a->poly.i32 OP b->poly.i64; a->type = I64; break;
+        case U64: a->poly.u64 = a->poly.i32 OP b->poly.u64; a->type = U64; break;
+        case F32: a->poly.f32 = a->poly.i32 OP b->poly.f32; a->type = F32; break;
+        case F64: a->poly.f64 = a->poly.i32 OP b->poly.f64; a->type = F64; break;
         default: BAIL; break;
     }
     break;
@@ -95,10 +95,10 @@ case U32:
         case U16: a->poly.u32 = a->poly.u32 OP b->poly.u16; a->type = U32; break;
         case I32: a->poly.u32 = a->poly.u32 OP b->poly.i32; a->type = U32; break;
         case U32: a->poly.u32 = a->poly.u32 OP b->poly.u32; a->type = U32; break;
-        case I64: a->poly.u32 = a->poly.u32 OP b->poly.i64; a->type = I64; break;
-        case U64: a->poly.u32 = a->poly.u32 OP b->poly.u64; a->type = U64; break;
-        case F32: a->poly.u32 = a->poly.u32 OP b->poly.f32; a->type = F32; break;
-        case F64: a->poly.u32 = a->poly.u32 OP b->poly.f64; a->type = F64; break;
+        case I64: a->poly.i64 = a->poly.u32 OP b->poly.i64; a->type = I64; break;
+        case U64: a->poly.u64 = a->poly.u32 OP b->poly.u64; a->type = U64; break;
+        case F32: a->poly.f32 = a->poly.u32 OP b->poly.f32; a->type = F32; break;
+        case F64: a->poly.f64 = a->poly.u32 OP b->poly.f64; a->type = F64; break;
         default: BAIL; break;
     }
     break;
@@ -112,9 +112,9 @@ case I64:
         case I32: a->poly.i64 = a->poly.i64 OP b->poly.i32; a->type = I64; break;
         case U32: a->poly.i64 = a->poly.i64 OP b->poly.u32; a->type = I64; break;
         case I64: a->poly.i64 = a->poly.i64 OP b->poly.i64; a->type = I64; break;
-        case U64: a->poly.i64 = a->poly.i64 OP b->poly.u64; a->type = U64; break;
-        case F32: a->poly.i64 = a->poly.i64 OP b->poly.f32; a->type = F32; break;
-        case F64: a->poly.i64 = a->poly.i64 OP b->poly.f64; a->type = F64; break;
+        case U64: a->poly.u64 = a->poly.i64 OP b->poly.u64; a->type = U64; break;
+        case F32: a->poly.f32 = a->poly.i64 OP b->poly.f32; a->type = F32; break;
+        case F64: a->poly.f64 = a->poly.i64 OP b->poly.f64; a->type = F64; break;
         default: BAIL; break;
     }
     break;
@@ -129,8 +129,8 @@ case U64:
         case U32: a->poly.u64 = a->poly.u64 OP b->poly.u32; a->type = U64; break;
         case I64: a->poly.u64 = a->poly.u64 OP b->poly.i64; a->type = U64; break;
         case U64: a->poly.u64 = a->poly.u64 OP b->poly.u64; a->type = U64; break;
-        case F32: a->poly.u64 = a->poly.u64 OP b->poly.f32; a->type = F32; break;
-        case F64: a->poly.u64 = a->poly.u64 OP b->poly.f64; a->type = F64; break;
+        case F32: a->poly.f32 = a->poly.u64 OP b->poly.f32; a->type = F32; break;
+        case F64: a->poly.f64 = a->poly.u64 OP b->poly.f64; a->type = F64; break;
         default: BAIL; break;
     }
     break;
@@ -146,7 +146,7 @@ case F32:
         case I64: a->poly.f32 = a->poly.f32 OP b->poly.i64; a->type = F32; break;
         case U64: a->poly.f32 = a->poly.f32 OP b->poly.u64; a->type = F32; break;
         case F32: a->poly.f32 = a->poly.f32 OP b->poly.f32; a->type = F32; break;
-        case F64: a->poly.f32 = a->poly.f32 OP b->poly.f64; a->type = F64; break;
+        case F64: a->poly.f64 = a->poly.f32 OP b->poly.f64; a->type = F64; break;
         default: BAIL; break;
     }
     break;
